Move attachment related stuff to attachment package
